// frontend/src/api/generated/models/ApartmentWithRooms.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Rental API
 * API for rental management
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { RoomDetails } from './RoomDetails';
import {
    RoomDetailsFromJSON,
    RoomDetailsFromJSONTyped,
    RoomDetailsToJSON,
} from './RoomDetails';

/**
 * 
 * @export
 * @interface ApartmentWithRooms
 */
export interface ApartmentWithRooms {
    /**
     * 
     * @type {number}
     * @memberof ApartmentWithRooms
     */
    apartmentId: number;
    /**
     * 
     * @type {string}
     * @memberof ApartmentWithRooms
     */
    apartmentName: string;
    /**
     * 
     * @type {Array<RoomDetails>}
     * @memberof ApartmentWithRooms
     */
    rooms: Array<RoomDetails>;
}

/**
 * Check if a given object implements the ApartmentWithRooms interface.
 */
export function instanceOfApartmentWithRooms(value: object): value is ApartmentWithRooms {
    if (!('apartmentId' in value) || value['apartmentId'] === undefined) return false;
    if (!('apartmentName' in value) || value['apartmentName'] === undefined) return false;
    if (!('rooms' in value) || value['rooms'] === undefined) return false;
    return true;
}

export function ApartmentWithRoomsFromJSON(json: any): ApartmentWithRooms {
    return ApartmentWithRoomsFromJSONTyped(json, false);
}

export function ApartmentWithRoomsFromJSONTyped(json: any, ignoreDiscriminator: boolean): ApartmentWithRooms {
    if (json == null) {
        return json;
    }
    return {
        
        'apartmentId': json['apartmentId'],
        'apartmentName': json['apartmentName'],
        'rooms': ((json['rooms'] as Array<any>).map(RoomDetailsFromJSON)),
    };
}

export function ApartmentWithRoomsToJSON(value?: ApartmentWithRooms | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'apartmentId': value['apartmentId'],
        'apartmentName': value['apartmentName'],
        'rooms': ((value['rooms'] as Array<any>).map(RoomDetailsToJSON)),
    };
}